+++ src/components/Profile/index.ts
@@ -1,0 +1,3 @@
+export { EditProfile } from "./EditProfile";
+export { ProfileDetails } from "./ProfileDetails";
+export { ProfileContainer } from "./ProfileContainer";
